requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.4.0
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.0
openai==1.35.3
openai-whisper @ git+https://github.com/openai/whisper.git@5979f03701209bb035a0a466f14131aeb1116cbb
outcome==1.3.0.post0
packaging==24.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PySocks==1.7.1
python-dotenv==1.0.1
regex==2024.9.11
requests==2.31.0
scipy==1.13.1
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.0
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.0
wsproto==1.2.0
